// frontend-angular/src/app/api/models/comment.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserPublic } from '../models/user-public';
export interface Comment {
  author: UserPublic;
  body: string;
  createdAt: string;

  /**
   * null until the comment's first edit (S8.1)
   */
  editedAt: string | null;
  id: number;
  ticketId: number;
}